refactor(core.grasshopper): reorganiation of classes and namespace update
