+++ src/SmartHopper.Core.Grasshopper/Utils/GHCanvasUtils.cs
@@ -1,7 +1,7 @@
-﻿/*
+/*
  * SmartHopper - AI-powered Grasshopper Plugin
  * Copyright (C) 2024 Marc Roca Musach
- * 
+ *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
@@ -14,7 +14,7 @@ using System;
 using System.Collections.Generic;
 using System.Drawing;
 
-namespace SmartHopper.Core.Grasshopper
+namespace SmartHopper.Core.Grasshopper.Utils
 {
     public class GHCanvasUtils
     {
@@ -89,3 +89,4 @@ namespace SmartHopper.Core.Grasshopper
         }
     }
 }
+
